requirements.txt
# Core framework
fastapi>=0.110.0
uvicorn[standard]>=0.27.1

# File handling & validation
python-multipart>=0.0.9

# Data validation
pydantic>=2.7.0

# OCR and File parsing (Step 2-3)
pytesseract>=0.3.10
Pillow>=10.2.0
PyMuPDF>=1.23.0

# AI (Step 5)
ollama>=0.1.6

# Testing
pytest>=8.1.1
httpx>=0.27.0
